up: remove some files
