+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/gookit/gcli/v3
 
-go 1.12
+go 1.13
 
 require (
 	github.com/fsnotify/fsnotify v1.4.9
